Breaking pixeldataset.py up into separate files
